requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# RyanairTicketReserve/main.py: 9
python-decouple 

# RyanairTicketReserve/main.py: 4,5,6,7,8
selenium == 3.141.0
